% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRmultinom.R
\name{AME}
\alias{AME}
\title{Average Marginal effect}
\usage{
AME(model, data, method = c("delta", "simulation"), K = 500, eps = 1e-07)
}
\arguments{
\item{model}{the model object of class \code{\link{rr_multinom}}.}

\item{data}{The data set on which to calculate the marginal effect.}

\item{method}{character specifying the method of estimation: delta or simulation.}

\item{K}{integer number indicating the number of iterations if method "simulation" is specified.}

\item{eps}{the value of the step used in calculating the marginal effect of numerical covariates.}
}
\value{
data frame with each row is an observations and columns are the marginal effect for all covariates on the estimated probabilities
of never, former and recent categories.
}
\description{
calculates average marginal effect (AME) along with its standard error
for all numerical and categorical covariates of the model object \code{\link{rr_multinom}}.
}
\details{
For numerical covariates, marginal effect is calculated using two sided numerical differentiation
adopted from \url{https://github.com/leeper/margins}.
}
\examples{
regression <- rr_multinom(formula = ~BodyBuild+Age, data = Doping,
                         response = c("Q9", "Q10"), p00 = 5/6, p11 = 5/6)
ME <- AME(model = regression, data = subset(Doping, BodyBuild == 0),
          method = "delta")
df <- ME$Age
}
\references{
Leeper TJ (2021). margins: Marginal Effects for Model Objects. R package version 0.3.26.
}
