% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRmultinom.R
\name{prev_bivar}
\alias{prev_bivar}
\title{Prevalence estimation for the bivariate model of the "ever"
and "last year" RR Questions}
\usage{
prev_bivar(Q1, Q2, data, p00, p11, seed = NULL)
}
\arguments{
\item{Q1}{name of the question "Have you ever...?".}

\item{Q2}{name of the question "In the last year, Did you ...?"}

\item{data}{data frame with individual records containing \code{Q1} and
\code{Q2} variables with values 1 denoting "yes" and 0 for "no" answer.}

\item{p00}{The conditional probability of observed "no" while
true response is "no".}

\item{p11}{The conditional probability of observed "yes" while
true response is "yes".}

\item{seed}{a single integer value. Default: \code{seed = NULL}. If you got \code{NAN}, you should specify
seeds.}
}
\value{
A list with the elements:
\describe{
\item{est}{data frame with point and interval estimates for the MLE.}
\item{gof}{Goodness-of-fit statistic for the bivariate model.}
\item{freqs}{data frame with the observed and estimated response frequencies.}
}
}
\description{
Point and interval estimates of the
Maximum Likelihood Estimator (MLE) for prevalence
of never, former and recent categories through the bivariate model of the "ever"
and "last year" randomized response (RR) questions.
}
\examples{
prev_bivar(Q1 = Q9, Q2 = Q10, data = Doping, p00 = 5/6, p11 = 5/6)

}
