% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRmultinom.R
\name{rr_multinom}
\alias{rr_multinom}
\title{Multinomial logistic RR regression model}
\usage{
rr_multinom(formula, data, response, p00, p11, iter = 500)
}
\arguments{
\item{formula}{a "\code{\link{formula}}" in the form
\code{~ terms}, where  \code{terms} describes the linear predictors.}

\item{data}{data frame with individual records containing the
\code{response} with values 1 for "yes", 0 for "no", and the predictors.}

\item{response}{character vector with the names of the response variables of the
two questions "Have you ever...?" and "in the last year, Did you ...?", in that order.}

\item{p00}{The conditional probability of observed "no" while
true response is "no".}

\item{p11}{The conditional probability of observed "yes" while
true response is "yes".}

\item{iter}{The maximum number of iterations of optimization method "Nelder-Mead". Default is 500.}
}
\value{
A list with the following objects:
\describe{
\item{formula}{formula used in the model specification}
\item{Call}{the matched call}
\item{coefs}{matrix of the regression coefficients}
\item{parmsig}{data frame with the parameter estimates and its signficance.}
\item{gof}{fit statistics including AIC and a likelihood-ratio
test against the intercept-only model.}
\item{fitted}{data frame with the fitted probabilities.}
\item{varmat}{variance-covariance matrix of the parameter estimates.}
\item{modeldata}{data frame with the covariates used in \code{formula}}
\item{D}{design matrix of the fitted model.}
}
}
\description{
Fits a multinomial logistic regression model to the bivariate analysis of the "ever"
and "last year" Questions to investigate potential covariates associated
with the sensitive characteristic.
}
\examples{
regression <- rr_multinom(formula = ~BodyBuild+Age, data = Doping,
response = c("Q9", "Q10"), p00 = 5/6, p11 = 5/6)

}
